orte mpi tests: fix a bunch of compiler warnings

+++ orte/test/mpi/ziatest.c
@@ -2,6 +2,7 @@
  *
  * Copyright (c) 2008 Los Alamos National Security, LLC.  All rights reserved.
  *
+ * Copyright (c) 2015 Cisco Systems, Inc.  All rights reserved.
  * $COPYRIGHT$
  * 
  * Additional copyrights may follow
@@ -15,6 +16,7 @@
 #include <unistd.h>
 #include <stdlib.h>
 #include <time.h>
+#include <sys/time.h>
 
 #include <mpi.h>
 
@@ -38,7 +40,8 @@ int main(int argc, char* argv[])
     gettimeofday(&tv, NULL);
     
     /* form the command */
-    asprintf(&cmd, "mpirun -npernode %d ./ziaprobe %d %d", nppn, tv.tv_sec, tv.tv_usec);
+    asprintf(&cmd, "mpirun -npernode %d ./ziaprobe %ld %d",
+             nppn, (long) tv.tv_sec, tv.tv_usec);
     
     /* execute it */
     system(cmd);
